Enemy now inflicts damage on Player via the CharacterColliderService. (#29)
WieldedWeapon is now initialized on CharacterBase.Start by searching for the Weapon component on children. This works for the MainPlayerCharacter as well because the Equipment attachment is done at Awake, so the initialization from EquipmentService has been removed for now as it is redundant, but will probably be readded in the future whenever we implement weapon switching.